Expose dialog's ID; extract merge_data and merge_aria into helper module
